FEM: unit tests, rename test tools to utils to distinguish between test modules
