+++ src/Mod/Fem/femtest/testccxtools.py
@@ -28,8 +28,8 @@ from femtools import ccxtools
 import FreeCAD
 import ObjectsFem
 import unittest
-from . import testtools
-from .testtools import fcc_print
+from . import utilstest as testtools
+from .utilstest import fcc_print
 
 
 class TestCcxTools(unittest.TestCase):
