+++ src/Mod/Fem/femtest/testcommon.py
@@ -26,8 +26,8 @@
 import FreeCAD
 import ObjectsFem
 import unittest
-from . import testtools
-from .testtools import fcc_print
+from . import utilstest as testtools
+from .utilstest import fcc_print
 
 
 class TestFemCommon(unittest.TestCase):
