+++ src/Mod/Fem/femtest/testmesh.py
@@ -26,8 +26,8 @@
 import Fem
 import FreeCAD
 import unittest
-from . import testtools
-from .testtools import fcc_print
+from . import utilstest as testtools
+from .utilstest import fcc_print
 
 
 class TestMeshCommon(unittest.TestCase):
